Move toX into toX.go
